Implementation of Course Unit Tests
10/12 unit tests passed

GET /courses/{course_id}
Unit Tests
1.0.0 - testFindCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.
Test Results: Passed

1.0.1 - testFindCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.
Test Results: Passed

GET /courses
Unit Tests
1.0.0 - testFindAllCourses
Purpose: Validate the functionality of the method.
Test Results: Passed

POST /courses
Unit Tests
1.0.0 - testCreateValidCourse
Purpose: Validate the functionality of the method when provided with a valid id.
Test Results:
	Argument(s) are different! Wanted:
courseRepo.save(
    ca.mcgill.ecse321.SportCenterManager.model.Course@6f18445b
);
-> at ca.mcgill.ecse321.SportCenterManager.service.CourseServiceTests.testCreateValidCourse(CourseServiceTests.java:98)
Actual invocations have different arguments:
courseRepo.findAll(

);
-> at ca.mcgill.ecse321.SportCenterManager.service.EventService.findAllCourses(EventService.java:31)
courseRepo.save(
    ca.mcgill.ecse321.SportCenterManager.model.Course@69bdcfbc
);
-> at ca.mcgill.ecse321.SportCenterManager.service.EventService.createCourse(EventService.java:86)

1.0.1 - testCreateCourseWithConflictingName
Purpose: Validate the functionality of the method when provided with a name that an existing course has.
Test Results: Passed

1.0.2 - testCreateCourseWithInvalidCostPerSession
Purpose: Validate the functionality of the method when provided with an invalid cost per session (less or equal to zero).
Test Results: Passed

PUT /courses/{course_id}
1.0.0 - testModifyCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.
Test Results: Passed

1.0.1 - testModifyCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.
Test Results: Passed

1.0.2 - testModifyCourseWithInvalidCostPerSession
Purpose: Validate the functionality of the method when provided with an invalid cost per session (less or equal to zero).
Test Results: Passed

DELETE /courses/{course_id}
Unit Tests
1.0.0 - testDeleteCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.
Test Results: Passed

1.0.1 - testDeleteCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.
Test Results: Passed

DELETE /courses/{course_id}/sessions
Unit Tests
1.0.0 - testDeleteAllSessionsCourse
Purpose: Validate the functionality of the method
Test Results:
org.mockito.exceptions.verification.TooManyActualInvocations:
sessionRepo.findAll();
Wanted 1 time:
-> at ca.mcgill.ecse321.SportCenterManager.service.CourseServiceTests.testDeleteAllSessionsOfCourse(CourseServiceTests.java:291)
But was 2 times:
-> at ca.mcgill.ecse321.SportCenterManager.service.EventService.findAllSessionsOfCourse(EventService.java:93)
-> at ca.mcgill.ecse321.SportCenterManager.service.EventService.findAllSessionsOfCourse(EventService.java:93)

+++ SportCenterManager-Backend/src/main/java/ca/mcgill/ecse321/SportCenterManager/model/Course.java
@@ -43,7 +43,7 @@ public class Course
     return wasSet;
   }
 
-  public boolean setCostPerSession(int aCostPerSession)
+  public boolean setCostPerSession(double aCostPerSession)
   {
     boolean wasSet = false;
     costPerSession = aCostPerSession;
